// backend/src/audit-log/audit-action.enum.ts
export enum AuditAction {
  AuthLogin = 'AUTH_LOGIN',
  AuthLogout = 'AUTH_LOGOUT',
  AuthRefresh = 'AUTH_REFRESH',
  AuthRefreshFail = 'AUTH_REFRESH_FAIL',
  UserCreate = 'USER_CREATE',
  UserUpdate = 'USER_UPDATE',
  UserDelete = 'USER_DELETE',
  UserRoleChange = 'USER_ROLE_CHANGE',
  CourseCreate = 'COURSE_CREATE',
  CourseUpdate = 'COURSE_UPDATE',
  CourseDelete = 'COURSE_DELETE',
  ContentCreate = 'CONTENT_CREATE',
  ContentUpdate = 'CONTENT_UPDATE',
  ContentDelete = 'CONTENT_DELETE',
  CourseEnroll = 'COURSE_ENROLL',
  CourseUnenroll = 'COURSE_UNENROLL',
  CourseFavorite = 'COURSE_FAVORITE',
  CourseUnfavorite = 'COURSE_UNFAVORITE',
}